# docs/swagger.yaml
definitions:
  apperrors.HTTPError:
    properties:
      code:
        type: integer
      details: {}
      message:
        type: string
    type: object
  user.AccessTokenResponse:
    properties:
      access_token:
        type: string
    type: object
  user.LoginRequest:
    properties:
      password:
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  user.LoginResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  user.SignUpRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      gender:
        enum:
        - male
        - female
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - age
    - email
    - gender
    - password
    - username
    type: object
  user.UpdateUserInput:
    properties:
      age:
        type: integer
      email:
        type: string
      gender:
        enum:
        - male
        - female
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
      weight:
        type: number
    type: object
  user.User:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
      weight:
        type: number
    type: object
info:
  contact: {}
paths:
  /api/users/{id}:
    get:
      description: Admin endpoint to fetch user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - admin
  /api/users/me:
    get:
      description: Returns the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Updates fields of the authenticated user's profile
      parameters:
      - description: Update user payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
  /auth/login:
    post:
      consumes:
      - application/json
      description: Validates credentials and returns access token + sets refresh token
        cookie
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      description: Revokes the refresh token and clears the cookie
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
      summary: Logout user
      tags:
      - auth
  /auth/refresh:
    post:
      description: Rotates refresh token and returns new access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.AccessTokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
      summary: Refresh access token
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: Sign up payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.HTTPError'
      summary: Register a new user
      tags:
      - auth
swagger: "2.0"